Add: Show control window over chart window when . . .
when in Single Screen mode.  Specifically, only in Single Screen Mode,
when the mouse is over the right side of the screen the control window
is not visible and the entire chart window is visible.  When the mouse
is over the left side of the screen the control window become visible.