Got Left Eye Visual Acuity Can Read  to work
This was a lot of heart ache.  I never could get the ValueConver to be
called.  So I reverted to subscribing to PropertyChanged from within the
view and doing a lot of if statements to figure out which tab is active.
For now it only works for Visual Acuity -- Left Eye.